dont provide stop suggestions when we dont have the pattern path data
This was causing issues when the pattern path api queries have not completed (ie when the bus api is slow). We need the pattern path data to get GPS to send to route planner and if we dont have it, it shows the stop as an option but does not provide any route options. This solution at least allows them to use the google suggestions which are fine.